requirements.txt
crewai
crewai-tools
langchain-groq
python-dotenv
litellm
chainlit
apscheduler
email-validator
fastapi-sso